sửa logic nghiệp vụ profile cho candidate
Add CV upload flow in apply modal (UploadCvModal endpoint, drag/drop UI, client validation, upload-before-apply) and extend apply VM to accept CvId and candidate contact fields. Introduce Combined_Seed.sql sample data and add multiple EF migrations (author_name, email_notifications_enabled mapping, SyncSchemaChanges scaffold) and model snapshot updates (email_notifications_enabled default true). Adjust interview insert scripts to remove recruiter_id, improve dropping unique constraint, and fix job search salary handling (DesiredSalary as decimal, no unit multiplication). Update connection strings and add sample CV file.

+++ ViewModels/Jobs/JobSearchFilterViewModel.cs
@@ -8,7 +8,7 @@ public class JobSearchFilterViewModel
     public string? Keyword { get; set; }
     public string? WorkingModel { get; set; }
     public string? ExperienceLevel { get; set; }
-    public int? DesiredSalary { get; set; }
+    public decimal? DesiredSalary { get; set; }
     // Quick-filter: lọc theo kỹ năng / thành phố / công ty
     public int? TechId { get; set; }
     public string? FilterLocation { get; set; }
